requirements.txt
Pillow~=9.3.0
numpy~=1.24.3